Refactor serde package
