+++ serde/testhelpers.go
@@ -1,0 +1,36 @@
+package serde
+
+import (
+	"fmt"
+	"reflect"
+	"runtime"
+	"testing"
+)
+
+type failFunc func(string, ...error)
+
+var maybeFail failFunc
+
+func initFailFunc(t *testing.T) failFunc {
+	tester := t
+	return func(msg string, errors ...error) {
+		for _, err := range errors {
+			if err != nil {
+				pc := make([]uintptr, 1)
+				runtime.Callers(2, pc)
+				caller := runtime.FuncForPC(pc[0])
+				_, line := caller.FileLine(caller.Entry())
+
+				tester.Fatalf("%s:%d failed: %s %s", caller.Name(), line, msg, err)
+			}
+		}
+	}
+}
+
+func expect(actual, expected interface{}) error {
+	if !reflect.DeepEqual(actual, expected) {
+		return fmt.Errorf("expected: %v, Actual: %v", expected, actual)
+	}
+
+	return nil
+}
